+++ .vscode/settings.json
@@ -1,0 +1,15 @@
+{
+  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  "[javascript]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  },
+  "typescript.tsdk": "node_modules/typescript/lib",
+  "typescript.enablePromptUseWorkspaceTsdk": true,
+  "workbench.colorCustomizations": {
+    // "titleBar.activeBackground": ""
+  },
+  "testing.automaticallyOpenPeekView": "never",
+  "jest.autoRun": {
+    "watch": false
+  }
+}
